\name{extendr-package}
\alias{extendr-package}
\alias{extendr}
\docType{package}
\title{
\packageTitle{extendr}
}
\description{
\packageDescription{extendr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{extendr}
\packageIndices{extendr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{extendr}

Maintainer: \packageMaintainer{extendr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
